+++ core/tests/all_tests.rs
@@ -12,6 +12,7 @@ mod default_scorer_tests;
 mod derive_attr_scorer_tests;
 mod fs_walk_tests;
 mod function_complexity_tests;
+mod generics_counter_tests;
 mod hidden_deps_counter_tests;
 mod macro_counter_tests;
 mod main_tests;